refactoring of project structure
