+++ js/modules/featureFilesModule.js
@@ -2,10 +2,10 @@ const fs = require('fs');
 const path = require('path');
 const Gherkin = require('@cucumber/gherkin');
 const Messages = require('@cucumber/messages');
-const gherkinDocumentToString = require('./gherkinUtils');
+const gherkinDocumentToString = require('./gherkinUtils.js');
 const { setNestedProperty, ensureDirectoryExistence } = require('./utils.js');
 
-let config = JSON.parse(fs.readFileSync(path.join(__dirname, 'config.json'), 'utf8'));
+let config = JSON.parse(fs.readFileSync(path.join(__dirname, '../../config.json'), 'utf8'));
 
 let featureFilesCopy = [];
 
@@ -103,7 +103,7 @@ function parseGherkinContent(content) {
 }
 
 function reset() {
-    config = JSON.parse(fs.readFileSync(path.join(__dirname, 'config.json'), 'utf8'));
+    config = JSON.parse(fs.readFileSync(path.join(__dirname, '../../config.json'), 'utf8'));
     const directoryPath = config.directoryPath;
     if (!directoryPath) {
         console.error('directoryPath non è definito nel file config.json');
