+++ views/pages/index.ejs
@@ -42,7 +42,7 @@
   <div id="wrapper">
     <!-- Sidebar -->
     <div id="sidebar-wrapper">
-      <%- include('sidebar.ejs') %>
+      <%- include('../partials/sidebar.ejs') %>
         <!-- End of Sidebar -->
     </div>
     <!-- Content Wrapper -->
@@ -52,7 +52,7 @@
         <nav class="navbar navbar-expand navbar-light bg-white mb-4 static-top shadow">
           <!-- Tags cloud-->
           <div class="container-fluid d-flex flex-wrap">
-            <%- include('tagCloud.ejs') %>
+            <%- include('../partials/tagCloud.ejs') %>
           </div>
         </nav>
         <!-- Begin Page Content -->
@@ -293,7 +293,7 @@
         </div>
       </div>
       <footer class="sticky-footer bg-white">
-        <%- include('footer.ejs') %>
+        <%- include('../partials/footer.ejs') %>
       </footer>
       <!-- End of Content Wrapper -->
     </div>
